cryptdrive/requirements.txt
Django==5.1.7
psycopg2==2.9.7
gunicorn==20.1.0

dotenv~=0.9.9
python-dotenv~=1.0.1
djangorestframework~=3.16.0